requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-docx==1.1.0
python-dotenv==1.0.1
langchain-text-splitters==0.0.1
mistralai>=1.0.0
sqlite-vec==0.1.6